use ERC721Keyed for developer registry
Also removes update. Update shouldn't be a valid operation, since this would need explicit approval. The right process flow for update would be to deactivate the developer and create a new one with updated values.

Also removes bot product creation. This will be done directly in the bot product registry, instead of proxied through this contract.